+++ HISTORY.rst
@@ -1,6 +1,7 @@
 XX (2017-XX-XX)
 ===============
 - Add Django-mathfilters.
+- Improve render of documents with no pages.
 
 2.3 (2017-06-08)
 ================